+++ examples/image.rs
@@ -55,10 +55,12 @@ fn main() -> Result<(), Box<dyn Error>> {
     let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
     let mut frame = Frame::new(0, 0, 400, 300, "");
 
-    let mut image = SharedImage::load(&PathBuf::from("screenshots/RustLogo.svg"))?;
+    let mut image = SharedImage::load(&PathBuf::from("screenshots/calc.jpg"))?;
     image.scale(200, 200, true, true);
 
-    frame.set_image(Some(image));
+    frame.set_deimage(Some(image));
+    let v = frame.deimage().unwrap();
+    frame.set_image(Some(v));
     
     // To remove an image
     // frame.set_image(Option::<SharedImage>::None);