+++ fltk-sys/cfltk/cfl_widget.cpp
@@ -215,6 +215,10 @@ void Fl_Widget_set_image(Fl_Widget *self, void *image) {
     LOCK(self->image(((Fl_Image *)image)))
 }
 
+void *Fl_Widget_image(const Fl_Widget *self) {
+    return (Fl_Image *)self->image();
+}
+
 void Fl_Widget_set_handler(Fl_Widget *self, custom_handler_callback cb, void *data) {
     LOCK(((Fl_Widget_Derived *)self)->set_handler_data(data);
          ((Fl_Widget_Derived *)self)->set_handler(cb);)
@@ -222,10 +226,6 @@ void Fl_Widget_set_handler(Fl_Widget *self, custom_handler_callback cb, void *da
 
 void Fl_Widget_set_trigger(Fl_Widget *self, int val) {
     LOCK(self->when(val);)
-}
-
-void *Fl_Widget_image(const Fl_Widget *self) {
-    return (Fl_Image *)self->image();
 }
 
 void Fl_Widget_set_draw(Fl_Widget *self, custom_draw_callback cb, void *data) {
@@ -329,4 +329,12 @@ void *Fl_Widget_as_window(Fl_Widget *self) {
 
 void *Fl_Widget_as_group(Fl_Widget *self) {
     return self->as_group();
+}
+
+void Fl_Widget_set_deimage(Fl_Widget *self, void *image) {
+    LOCK(self->deimage(((Fl_Image *)image)))
+}
+
+void *Fl_Widget_deimage(const Fl_Widget *self) {
+    return (Fl_Image *)self->deimage();
 }