+++ fltk-sys/src/browser.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Browser {
@@ -463,6 +469,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Browser_as_group(self_: *mut Fl_Browser) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Browser_set_deimage(arg1: *mut Fl_Browser, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Browser_deimage(arg1: *const Fl_Browser) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Browser_remove(arg1: *mut Fl_Browser, line: ::std::os::raw::c_int);
@@ -800,6 +812,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Hold_Browser_as_group(self_: *mut Fl_Hold_Browser) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Hold_Browser_set_deimage(
+        arg1: *mut Fl_Hold_Browser,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hold_Browser_deimage(arg1: *const Fl_Hold_Browser) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Hold_Browser_remove(arg1: *mut Fl_Hold_Browser, line: ::std::os::raw::c_int);
@@ -1171,6 +1192,16 @@ extern "C" {
         -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Select_Browser_set_deimage(
+        arg1: *mut Fl_Select_Browser,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Select_Browser_deimage(arg1: *const Fl_Select_Browser)
+        -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Select_Browser_remove(arg1: *mut Fl_Select_Browser, line: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -1526,6 +1557,15 @@ extern "C" {
     pub fn Fl_Multi_Browser_as_group(self_: *mut Fl_Multi_Browser) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Multi_Browser_set_deimage(
+        arg1: *mut Fl_Multi_Browser,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Multi_Browser_deimage(arg1: *const Fl_Multi_Browser) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Multi_Browser_remove(arg1: *mut Fl_Multi_Browser, line: ::std::os::raw::c_int);
 }
 extern "C" {
@@ -1867,6 +1907,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_File_Browser_as_group(self_: *mut Fl_File_Browser) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_File_Browser_set_deimage(
+        arg1: *mut Fl_File_Browser,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_File_Browser_deimage(arg1: *const Fl_File_Browser) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_File_Browser_remove(arg1: *mut Fl_File_Browser, line: ::std::os::raw::c_int);