+++ fltk-sys/src/button.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Button {
@@ -463,6 +469,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Button_as_group(self_: *mut Fl_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Button_set_deimage(arg1: *mut Fl_Button, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Button_deimage(arg1: *const Fl_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Button_shortcut(self_: *const Fl_Button) -> ::std::os::raw::c_int;
@@ -725,6 +737,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Check_Button_as_group(self_: *mut Fl_Check_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Check_Button_set_deimage(
+        arg1: *mut Fl_Check_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Check_Button_deimage(arg1: *const Fl_Check_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Check_Button_is_checked(arg1: *mut Fl_Check_Button) -> ::std::os::raw::c_int;
@@ -996,6 +1017,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Radio_Button_as_group(self_: *mut Fl_Radio_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Radio_Button_set_deimage(
+        arg1: *mut Fl_Radio_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Radio_Button_deimage(arg1: *const Fl_Radio_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Radio_Button_is_toggled(arg1: *mut Fl_Radio_Button) -> ::std::os::raw::c_int;
@@ -1281,6 +1311,15 @@ extern "C" {
     pub fn Fl_Toggle_Button_as_group(self_: *mut Fl_Toggle_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Toggle_Button_set_deimage(
+        arg1: *mut Fl_Toggle_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Toggle_Button_deimage(arg1: *const Fl_Toggle_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Toggle_Button_is_toggled(arg1: *mut Fl_Toggle_Button) -> ::std::os::raw::c_int;
 }
 extern "C" {
@@ -1550,6 +1589,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Round_Button_as_group(self_: *mut Fl_Round_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Round_Button_set_deimage(
+        arg1: *mut Fl_Round_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Round_Button_deimage(arg1: *const Fl_Round_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Round_Button_is_toggled(arg1: *mut Fl_Round_Button) -> ::std::os::raw::c_int;
@@ -1893,6 +1941,17 @@ extern "C" {
 extern "C" {
     pub fn Fl_Radio_Round_Button_as_group(
         self_: *mut Fl_Radio_Round_Button,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_set_deimage(
+        arg1: *mut Fl_Radio_Round_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_deimage(
+        arg1: *const Fl_Radio_Round_Button,
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
@@ -2250,6 +2309,17 @@ extern "C" {
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Radio_Light_Button_set_deimage(
+        arg1: *mut Fl_Radio_Light_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Radio_Light_Button_deimage(
+        arg1: *const Fl_Radio_Light_Button,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Radio_Light_Button_is_toggled(
         arg1: *mut Fl_Radio_Light_Button,
     ) -> ::std::os::raw::c_int;
@@ -2529,6 +2599,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Light_Button_as_group(self_: *mut Fl_Light_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Light_Button_set_deimage(
+        arg1: *mut Fl_Light_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Light_Button_deimage(arg1: *const Fl_Light_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Light_Button_is_on(arg1: *mut Fl_Light_Button) -> ::std::os::raw::c_int;
@@ -2814,6 +2893,15 @@ extern "C" {
     pub fn Fl_Repeat_Button_as_group(self_: *mut Fl_Repeat_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Repeat_Button_set_deimage(
+        arg1: *mut Fl_Repeat_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Repeat_Button_deimage(arg1: *const Fl_Repeat_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Repeat_Button_shortcut(self_: *const Fl_Repeat_Button) -> ::std::os::raw::c_int;
 }
 extern "C" {
@@ -3089,6 +3177,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Return_Button_as_group(self_: *mut Fl_Return_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Return_Button_set_deimage(
+        arg1: *mut Fl_Return_Button,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Return_Button_deimage(arg1: *const Fl_Return_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Return_Button_shortcut(self_: *const Fl_Return_Button) -> ::std::os::raw::c_int;