+++ fltk-sys/src/frame.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Box {
@@ -463,4 +469,10 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Box_as_group(self_: *mut Fl_Box) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Box_set_deimage(arg1: *mut Fl_Box, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Box_deimage(arg1: *const Fl_Box) -> *mut ::std::os::raw::c_void;
 }