+++ fltk-sys/src/input.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Input {
@@ -463,6 +469,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Input_as_group(self_: *mut Fl_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Input_set_deimage(arg1: *mut Fl_Input, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Input_deimage(arg1: *const Fl_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Input_set_value(
@@ -788,6 +800,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Int_Input_as_group(self_: *mut Fl_Int_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Int_Input_set_deimage(arg1: *mut Fl_Int_Input, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Int_Input_deimage(arg1: *const Fl_Int_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Int_Input_set_value(
@@ -1128,6 +1146,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Float_Input_as_group(self_: *mut Fl_Float_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Float_Input_set_deimage(arg1: *mut Fl_Float_Input, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Float_Input_deimage(arg1: *const Fl_Float_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Float_Input_set_value(
@@ -1523,6 +1547,17 @@ extern "C" {
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Multiline_Input_set_deimage(
+        arg1: *mut Fl_Multiline_Input,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Multiline_Input_deimage(
+        arg1: *const Fl_Multiline_Input,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Multiline_Input_set_value(
         arg1: *mut Fl_Multiline_Input,
         arg2: *const ::std::os::raw::c_char,
@@ -1888,6 +1923,15 @@ extern "C" {
     pub fn Fl_Secret_Input_as_group(self_: *mut Fl_Secret_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Secret_Input_set_deimage(
+        arg1: *mut Fl_Secret_Input,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Secret_Input_deimage(arg1: *const Fl_Secret_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Secret_Input_set_value(
         arg1: *mut Fl_Secret_Input,
         arg2: *const ::std::os::raw::c_char,
@@ -2217,6 +2261,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_File_Input_as_group(self_: *mut Fl_File_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_File_Input_set_deimage(arg1: *mut Fl_File_Input, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_File_Input_deimage(arg1: *const Fl_File_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_File_Input_set_value(