+++ fltk-sys/src/menu.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Menu_Item {
@@ -468,6 +474,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Menu_Bar_as_group(self_: *mut Fl_Menu_Bar) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Menu_Bar_set_deimage(arg1: *mut Fl_Menu_Bar, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Menu_Bar_deimage(arg1: *const Fl_Menu_Bar) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Menu_Bar_add(
@@ -826,6 +838,12 @@ extern "C" {
     pub fn Fl_Menu_Button_as_group(self_: *mut Fl_Menu_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Menu_Button_set_deimage(arg1: *mut Fl_Menu_Button, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Menu_Button_deimage(arg1: *const Fl_Menu_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Menu_Button_add(
         arg1: *mut Fl_Menu_Button,
         name: *const ::std::os::raw::c_char,
@@ -1165,6 +1183,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Choice_as_group(self_: *mut Fl_Choice) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Choice_set_deimage(arg1: *mut Fl_Choice, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Choice_deimage(arg1: *const Fl_Choice) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Choice_add(