+++ fltk-sys/src/text.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 pub type Fl_Text_Modify_Cb = ::std::option::Option<
     unsafe extern "C" fn(
         pos: ::std::os::raw::c_int,
@@ -688,6 +694,15 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Display_as_group(self_: *mut Fl_Text_Display) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Text_Display_set_deimage(
+        arg1: *mut Fl_Text_Display,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Text_Display_deimage(arg1: *const Fl_Text_Display) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Text_Display_init(arg1: *mut Fl_Text_Display);
@@ -1180,6 +1195,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Editor_as_group(self_: *mut Fl_Text_Editor) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Text_Editor_set_deimage(arg1: *mut Fl_Text_Editor, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Text_Editor_deimage(arg1: *const Fl_Text_Editor) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Text_Editor_init(arg1: *mut Fl_Text_Editor);
@@ -1731,6 +1752,17 @@ extern "C" {
 extern "C" {
     pub fn Fl_Simple_Terminal_as_group(
         self_: *mut Fl_Simple_Terminal,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_set_deimage(
+        arg1: *mut Fl_Simple_Terminal,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_deimage(
+        arg1: *const Fl_Simple_Terminal,
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {