+++ fltk-sys/src/valuator.rs
@@ -241,6 +241,12 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
 }
+extern "C" {
+    pub fn Fl_Widget_set_deimage(arg1: *mut Fl_Widget, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Widget_deimage(arg1: *const Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Slider {
@@ -463,6 +469,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Slider_as_group(self_: *mut Fl_Slider) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Slider_set_deimage(arg1: *mut Fl_Slider, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Slider_deimage(arg1: *const Fl_Slider) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Slider_set_bounds(arg1: *mut Fl_Slider, a: f64, b: f64);
@@ -755,6 +767,12 @@ extern "C" {
     pub fn Fl_Nice_Slider_as_group(self_: *mut Fl_Nice_Slider) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Nice_Slider_set_deimage(arg1: *mut Fl_Nice_Slider, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Nice_Slider_deimage(arg1: *const Fl_Nice_Slider) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Nice_Slider_set_bounds(arg1: *mut Fl_Nice_Slider, a: f64, b: f64);
 }
 extern "C" {
@@ -1028,6 +1046,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Counter_as_group(self_: *mut Fl_Counter) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Counter_set_deimage(arg1: *mut Fl_Counter, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Counter_deimage(arg1: *const Fl_Counter) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Counter_set_bounds(arg1: *mut Fl_Counter, a: f64, b: f64);
@@ -1305,6 +1329,12 @@ extern "C" {
     pub fn Fl_Dial_as_group(self_: *mut Fl_Dial) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Dial_set_deimage(arg1: *mut Fl_Dial, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Dial_deimage(arg1: *const Fl_Dial) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Dial_set_bounds(arg1: *mut Fl_Dial, a: f64, b: f64);
 }
 extern "C" {
@@ -1577,6 +1607,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Line_Dial_as_group(self_: *mut Fl_Line_Dial) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Line_Dial_set_deimage(arg1: *mut Fl_Line_Dial, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Line_Dial_deimage(arg1: *const Fl_Line_Dial) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Line_Dial_set_bounds(arg1: *mut Fl_Line_Dial, a: f64, b: f64);
@@ -1854,6 +1890,12 @@ extern "C" {
     pub fn Fl_Roller_as_group(self_: *mut Fl_Roller) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Roller_set_deimage(arg1: *mut Fl_Roller, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Roller_deimage(arg1: *const Fl_Roller) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Roller_set_bounds(arg1: *mut Fl_Roller, a: f64, b: f64);
 }
 extern "C" {
@@ -2127,6 +2169,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Scrollbar_as_group(self_: *mut Fl_Scrollbar) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Scrollbar_set_deimage(arg1: *mut Fl_Scrollbar, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Scrollbar_deimage(arg1: *const Fl_Scrollbar) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Scrollbar_set_bounds(arg1: *mut Fl_Scrollbar, a: f64, b: f64);
@@ -2428,6 +2476,15 @@ extern "C" {
     pub fn Fl_Value_Slider_as_group(self_: *mut Fl_Value_Slider) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Value_Slider_set_deimage(
+        arg1: *mut Fl_Value_Slider,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Value_Slider_deimage(arg1: *const Fl_Value_Slider) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Value_Slider_set_bounds(arg1: *mut Fl_Value_Slider, a: f64, b: f64);
 }
 extern "C" {
@@ -2704,6 +2761,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Adjuster_as_group(self_: *mut Fl_Adjuster) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Adjuster_set_deimage(arg1: *mut Fl_Adjuster, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Adjuster_deimage(arg1: *const Fl_Adjuster) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Adjuster_set_bounds(arg1: *mut Fl_Adjuster, a: f64, b: f64);
@@ -2997,6 +3060,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Value_Input_as_group(self_: *mut Fl_Value_Input) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Value_Input_set_deimage(arg1: *mut Fl_Value_Input, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Value_Input_deimage(arg1: *const Fl_Value_Input) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Value_Input_set_bounds(arg1: *mut Fl_Value_Input, a: f64, b: f64);
@@ -3298,6 +3367,15 @@ extern "C" {
     pub fn Fl_Value_Output_as_group(self_: *mut Fl_Value_Output) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Value_Output_set_deimage(
+        arg1: *mut Fl_Value_Output,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Value_Output_deimage(arg1: *const Fl_Value_Output) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Value_Output_set_bounds(arg1: *mut Fl_Value_Output, a: f64, b: f64);
 }
 extern "C" {
@@ -3594,6 +3672,12 @@ extern "C" {
     pub fn Fl_Fill_Slider_as_group(self_: *mut Fl_Fill_Slider) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
+    pub fn Fl_Fill_Slider_set_deimage(arg1: *mut Fl_Fill_Slider, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Fill_Slider_deimage(arg1: *const Fl_Fill_Slider) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Fill_Slider_set_bounds(arg1: *mut Fl_Fill_Slider, a: f64, b: f64);
 }
 extern "C" {
@@ -3870,6 +3954,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Fill_Dial_as_group(self_: *mut Fl_Fill_Dial) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Fill_Dial_set_deimage(arg1: *mut Fl_Fill_Dial, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Fill_Dial_deimage(arg1: *const Fl_Fill_Dial) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Fill_Dial_set_bounds(arg1: *mut Fl_Fill_Dial, a: f64, b: f64);
@@ -4154,6 +4244,12 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Hor_Slider_as_group(self_: *mut Fl_Hor_Slider) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Hor_Slider_set_deimage(arg1: *mut Fl_Hor_Slider, arg2: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Hor_Slider_deimage(arg1: *const Fl_Hor_Slider) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Hor_Slider_set_bounds(arg1: *mut Fl_Hor_Slider, a: f64, b: f64);
@@ -4499,6 +4595,17 @@ extern "C" {
 extern "C" {
     pub fn Fl_Hor_Fill_Slider_as_group(
         self_: *mut Fl_Hor_Fill_Slider,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Hor_Fill_Slider_set_deimage(
+        arg1: *mut Fl_Hor_Fill_Slider,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Fill_Slider_deimage(
+        arg1: *const Fl_Hor_Fill_Slider,
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
@@ -4855,6 +4962,17 @@ extern "C" {
 extern "C" {
     pub fn Fl_Hor_Nice_Slider_as_group(
         self_: *mut Fl_Hor_Nice_Slider,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Hor_Nice_Slider_set_deimage(
+        arg1: *mut Fl_Hor_Nice_Slider,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Nice_Slider_deimage(
+        arg1: *const Fl_Hor_Nice_Slider,
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
@@ -5220,6 +5338,17 @@ extern "C" {
 extern "C" {
     pub fn Fl_Hor_Value_Slider_as_group(
         self_: *mut Fl_Hor_Value_Slider,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Hor_Value_Slider_set_deimage(
+        arg1: *mut Fl_Hor_Value_Slider,
+        arg2: *mut ::std::os::raw::c_void,
+    );
+}
+extern "C" {
+    pub fn Fl_Hor_Value_Slider_deimage(
+        arg1: *const Fl_Hor_Value_Slider,
     ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {