+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 - Change interface for working with images. Now widgets take an Option<ImageExt>. This allows unsetting images.
 - Remove unnecessary copy when setting a widget image.
+- Add deimage and set_deimage methods to WidgetExt.
 
 ## [0.6.9] - 2020-06-20
 ### Changes